backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver
redis==5.0.1

# Vector DB
weaviate-client==4.4.0

# Celery
celery==5.3.6
celery[redis]==5.3.6

# Crawling
requests==2.31.0
beautifulsoup4==4.12.3
trafilatura==1.7.0
lxml==5.1.0

# Embeddings & LLM
openai==1.10.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0